+++ src/components/Tweet.vue
@@ -1,5 +1,5 @@
 <template lang='pug'>
-a.twitter-share-button(:data-text='title', :data-url='url', data-size='large', href='https://twitter.com/intent/tweet') Tweet
+a.twitter-share-button(:data-text='"Join me for " + title', :data-url='url', data-size='large', href='https://twitter.com/intent/tweet') Tweet
 </template>
 
 <script lang="ts">
@@ -11,10 +11,3 @@ export default class Tweet extends Vue {
   @Prop({ default: 'https://www.schoolhouseyoga.com/workshops' }) private url!: string;
 }
 </script>
-
-<style scoped lang="scss">
-.twitter-share-button {
-  width: 78px;
-  height: 26px;
-}
-</style>