+++ src/views/Workshops.vue
@@ -10,10 +10,24 @@
 import { Component, Vue, Prop } from 'vue-property-decorator';
 import Tweet from '@/components/Tweet.vue';
 
+declare let window: any; // suppress window.twttr error
+
 @Component({
   components: {
     Tweet,
   },
 })
-export default class Workshops extends Vue {}
+export default class Workshops extends Vue {
+  private mounted() {
+    window.twttr.widgets.load(); // call whenever using Tweet component
+  }
+
+}
 </script>
+
+<style lang="scss">
+.twitter-share-button {
+  width: 78px;
+  height: 26px;
+}
+</style>