+++ CHANGELOG.md
@@ -5,4 +5,6 @@
 * [High] Create UserManager component for Admin and SHYnet views
 * [High] Create components for Admin tabs (Dashboard, UserManager, ScheduleManager, AnnouncementManager, OrderManager, ProductManager, ClassManager, WorkshopManager, LocationManager)
 * [Medium] Create JSONLd component
-* [Low] Create Tweet component or use 3rd party
+
+<a name="1.0.0"></a>
+* Created Tweet component or use 3rd party